refactor(test): Move and update bootstrap utility for e2e tests
- Renamed `test.ts` to `bootstrap.ts` and moved it from `src/app/utils` to `test/utils`.
- Updated import paths across all e2e spec files to use the new bootstrap location.
- Improved `closeTestApp` to drop all collections instead of the entire database for cleaner test teardown.

Helps isolate test utilities and clarifies separation between application and test code.
