+++ packages/quiz-service/src/game/services/game.repository.e2e-spec.ts
@@ -3,7 +3,7 @@ import { getModelToken } from '@nestjs/mongoose'
 import { GameMode, GameStatus } from '@quiz/common'
 import { v4 as uuidv4 } from 'uuid'
 
-import { closeTestApp, createTestApp } from '../../app/utils/test'
+import { closeTestApp, createTestApp } from '../../../test/utils/bootstrap'
 import { Quiz } from '../../quiz/services/models/schemas'
 
 import { GameRepository } from './game.repository'
